+++ src/Client/Game/Objects/Player.cs
@@ -220,7 +220,7 @@ namespace Client
                 return canMove;
             }
 
-            var remaining = (int) (Player.Instance[GameState.MyIndex].DeathTimer - General.GetTickCount()) / 1000;
+            var remaining = (Player.Instance[GameState.MyIndex].DeathTimer - General.GetTickCount()) / 1000;
             if (remaining < 0) remaining = 0;
 
             if (remaining > 0)
